+++ backend/src/HorseRacing.Application/Features/OfficiatingAndResults/DTOs/AssignRefereeRequest.cs
@@ -2,5 +2,5 @@ namespace HorseRacing.Application.Features.OfficiatingAndResults.DTOs;
 
 public class AssignRefereeRequest
 {
-    public long RefereeId { get; set; }
+    public int RefereeId { get; set; }
 }
