+++ backend/src/HorseRacing.Application/Features/OfficiatingAndResults/Interfaces/IRefereeAssignmentRepository.cs
@@ -8,8 +8,8 @@ namespace HorseRacing.Application.Features.OfficiatingAndResults.Interfaces;
 public interface IRefereeAssignmentRepository
 {
     Task<Race?> GetRaceByIdAsync(long raceId);
-    Task<RefereeProfile?> GetRefereeProfileByIdAsync(long refereeId);
-    Task<RaceRefereeAssignment?> GetAssignmentAsync(long raceId, long refereeId);
+    Task<RefereeProfile?> GetRefereeProfileByIdAsync(int refereeId);
+    Task<RaceRefereeAssignment?> GetAssignmentAsync(long raceId, int refereeId);
     Task<List<RaceRefereeAssignment>> GetAssignmentsForRaceAsync(long raceId);
     Task AddAssignmentAsync(RaceRefereeAssignment assignment);
     void RemoveAssignment(RaceRefereeAssignment assignment);
