+++ backend/src/HorseRacing.Application/Features/OfficiatingAndResults/Interfaces/IRefereeAssignmentService.cs
@@ -8,5 +8,5 @@ public interface IRefereeAssignmentService
 {
     Task<RaceRefereeResponse> AssignRefereeAsync(long raceId, AssignRefereeRequest request);
     Task<List<RaceRefereeResponse>> GetAssignedRefereesAsync(long raceId);
-    Task RemoveRefereeAssignmentAsync(long raceId, long refereeId);
+    Task RemoveRefereeAssignmentAsync(long raceId, int refereeId);
 }
