+++ backend/src/HorseRacing.Domain/Entities/Users/RefereeProfile.cs
@@ -2,7 +2,7 @@ namespace HorseRacing.Domain.Entities;
 
 public class RefereeProfile
 {
-    public long RefereeId { get; set; }
+    public int RefereeId { get; set; }
     public int UserId { get; set; }
     public AppUser? User { get; set; }
     public string LicenseNumber { get; set; } = string.Empty;
