+++ backend/src/HorseRacing.Infrastructure/Migrations/AppDbContextModelSnapshot.cs
@@ -283,11 +283,11 @@ namespace HorseRacing.Infrastructure.Migrations
 
             modelBuilder.Entity("HorseRacing.Domain.Entities.JockeyProfile", b =>
                 {
-                    b.Property<long>("JockeyId")
+                    b.Property<int>("JockeyId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("JockeyId"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("JockeyId"));
 
                     b.Property<int>("ExperienceYears")
                         .HasColumnType("int");
@@ -374,8 +374,8 @@ namespace HorseRacing.Infrastructure.Migrations
                         .HasPrecision(10, 2)
                         .HasColumnType("decimal(10,2)");
 
-                    b.Property<long?>("JockeyId")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("JockeyId")
+                        .HasColumnType("int");
 
                     b.Property<int>("LaneNo")
                         .HasColumnType("int");
@@ -457,11 +457,11 @@ namespace HorseRacing.Infrastructure.Migrations
 
             modelBuilder.Entity("HorseRacing.Domain.Entities.RefereeProfile", b =>
                 {
-                    b.Property<long>("RefereeId")
+                    b.Property<int>("RefereeId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("RefereeId"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("RefereeId"));
 
                     b.Property<int>("ExperienceYears")
                         .HasColumnType("int");
@@ -682,8 +682,8 @@ namespace HorseRacing.Infrastructure.Migrations
                     b.Property<long>("RaceId")
                         .HasColumnType("bigint");
 
-                    b.Property<long>("RefereeId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("RefereeId")
+                        .HasColumnType("int");
 
                     b.Property<string>("Status")
                         .IsRequired()
